refactor(tamanu): extract shared readiness-probe module
Move the HTTP readiness-probe helpers (http_client, parse_duration,
probe_until_ready, probe_url, probe_once) out of restart.rs into a new
probe.rs, and extract the probe-URL construction (container IP for
systemd, pm2 PORT for pm2) into a shared instance_probe_url. restart's
probe_instance now calls into the shared module; no behaviour change.
